start filling out base models with schema described in README and main example
